[27] was created file Moc and CardTypes.
