+++ src/Types/CardTypes.ts
@@ -1,0 +1,8 @@
+export type CardsType = OneCardType[]
+export type OneCardType = {
+    id: number
+    title: string
+    userName: string
+    date: string
+    status: string | null
+}
